moved testing methods inside Tests/test_KDTree.py
